Add files via upload
In this, a Decision Tree model is developed to classify two types of raisins (Kecimen and Besni) based on various features like area, perimeter, and axis lengths. The notebook covers feature analysis, model training, and evaluation using confusion matrix and classification report. It also provides recommendations for improving the model, such as pruning and scaling.